// namespaces/clients/apps/clients/src/notes-tab/enums/DraftJobFormFields.ts
export enum DraftJobFormFields {
  BudgetDetails = 'budgetDetails',
  Commitment = 'commitment',
  Description = 'description',
  EstimatedLength = 'estimatedLength',
  HasPreferredHours = 'hasPreferredHours',
  HoursOverlap = 'hoursOverlap',
  Industries = 'industries',
  MaxHourlyRate = 'maxHourlyRate',
  ProjectSpecCompleteness = 'projectSpecCompleteness',
  ProjectTeamInvolved = 'projectTeamInvolved',
  ProjectType = 'projectType',
  Skills = 'skills',
  StartDate = 'startDate',
  TalentCount = 'talentCount',
  TimeZoneName = 'timeZoneName',
  Title = 'title',
  VerticalId = 'verticalId',
  WorkingTimeFrom = 'workingTimeFrom',
  WorkingTimeTo = 'workingTimeTo'
}